+++ src/main/java/com/telepathicgrunt/the_bumblezone/blocks/blockentities/HoneyCocoonBlockEntity.java
@@ -2,6 +2,7 @@ package com.telepathicgrunt.the_bumblezone.blocks.blockentities;
 
 import com.telepathicgrunt.the_bumblezone.blocks.HoneyCocoon;
 import com.telepathicgrunt.the_bumblezone.modinit.BzBlockEntities;
+import com.telepathicgrunt.the_bumblezone.modinit.BzBlocks;
 import com.telepathicgrunt.the_bumblezone.modinit.BzFluids;
 import com.telepathicgrunt.the_bumblezone.modinit.BzMenuTypes;
 import com.telepathicgrunt.the_bumblezone.screens.StrictChestMenu;
@@ -122,7 +123,7 @@ public class HoneyCocoonBlockEntity extends RandomizableContainerBlockEntity imp
 
         if (this.level != null) {
             BlockState blockState = this.level.getBlockState(this.worldPosition);
-            if (blockState.getValue(HoneyCocoon.WATERLOGGED)) {
+            if (blockState.is(BzBlocks.HONEY_COCOON.get()) && blockState.getValue(HoneyCocoon.WATERLOGGED)) {
                 this.level.scheduleTick(this.worldPosition, BzFluids.SUGAR_WATER_FLUID.get(), BzFluids.SUGAR_WATER_FLUID.get().getTickDelay(this.level));
                 this.level.scheduleTick(this.worldPosition, blockState.getBlock(), HoneyCocoon.waterDropDelay);
             }